+++ tests/testthat/test_powersetclassifiers.R
@@ -3,6 +3,7 @@ train <- toyml
 test <- toyml$dataset[10:40, toyml$attributesIndexes]
 
 predictionTest <- function (model) {
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(123)
   pred <- predict(model, test)
   expect_is(pred, "mlresult")
@@ -11,6 +12,7 @@ predictionTest <- function (model) {
   expect_equal(colnames(pred), rownames(toyml$labels))
   expect_equal(rownames(pred), rownames(test))
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(123)
   pred1 <- predict(model, test, prob = FALSE)
   expect_is(pred1, "mlresult")