+++ tests/testthat/test_seed.R
@@ -9,9 +9,11 @@ testTrainMethod <- function (method, name, ...) {
   }
 
   m1 <- method(smalltoyml, "SVM", ..., scale=FALSE, seed=123)
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(321)
   m2 <- method(smalltoyml, "SVM", ..., scale=FALSE, cores=1, seed=123)
   m3 <- method(smalltoyml, "SVM", ..., scale=FALSE, cores=2, seed=123)
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(345)
   m4 <- method(smalltoyml, "SVM", ..., scale=FALSE, cores=2, seed=123)
   m5 <- method(smalltoyml, "SVM", ..., scale=FALSE, cores=2, seed=NULL)
@@ -34,15 +36,18 @@ testPredictMethod <- function (method, name, ...) {
   model <- method(smalltoyml, "RANDOM", ...)
   r1 <- predict(model, smalltoyml, seed=123)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(321)
   r2 <- predict(model, smalltoyml, cores=1, seed=123)
   r3 <- predict(model, smalltoyml, cores=2, seed=123)
   v1 <- stats::runif(3)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(345)
   r4 <- predict(model, smalltoyml, cores=2, seed=123)
   r5 <- predict(model, smalltoyml, cores=2, seed=NULL)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(321)
   v2 <- stats::runif(3)
 
@@ -58,21 +63,26 @@ testSeedEffectMethod <- function (method, name, ...) {
     skip(name)
     return(FALSE)
   }
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(21)
   method(smalltoyml, "SVM", scale=FALSE, ...)
   v1 <- stats::runif(5)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(21)
   method(smalltoyml, "SVM", scale=FALSE, ..., seed=123)
   v2 <- stats::runif(5)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(21)
   method(smalltoyml, "SVM", scale=FALSE, ..., cores=2, seed=123)
   v3 <- stats::runif(5)
 
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(21)
   method(smalltoyml, "SVM", scale=FALSE, ..., cores=2, seed=321)
   v4 <- stats::runif(5)
+  suppressWarnings(RNGversion("3.5.0"))
   set.seed(22)
   method(smalltoyml, "SVM", scale=FALSE, ..., cores=2, seed=123)
   v5 <- stats::runif(5)